Use Go 1.25 synctest api in tests
