+++ lib/integrations/awsra/profile_syncer_test.go
@@ -1,5 +1,3 @@
-//go:build go1.24 && enablesynctest
-
 /*
  * Teleport
  * Copyright (C) 2025  Gravitational, Inc.
@@ -23,7 +21,6 @@ package awsra
 import (
 	"context"
 	"testing"
-	"testing/synctest"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -38,19 +35,8 @@ import (
 	"github.com/gravitational/teleport/lib/integrations/awsra/createsession"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
 	"github.com/gravitational/teleport/lib/utils/log/logtest"
+	"github.com/gravitational/teleport/lib/utils/testutils/synctest"
 )
-
-/*
-This file uses the experimental testing/synctest package introduced with Go 1.24:
-
-    https://go.dev/blog/synctest
-
-When editing this file, you should set GOEXPERIMENT=synctest for your editor/LSP
-to ensure that the language server doesn't fail to recognize the package.
-
-This file is also protected by a build tag to ensure that `go test` doesn't fail
-for users who haven't set the environment variable.
-*/
 
 func TestProfileSyncerTestAndSetDefaults(t *testing.T) {
 	keyStoreManager, err := keystore.NewManager(t.Context(), &servicecfg.KeystoreConfig{}, &keystore.Options{
@@ -242,18 +228,16 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: exampleProfileTags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
-			require.Len(t, serverClient.appServers, 0)
+			require.Empty(t, serverClient.appServers)
 		})
 	})
 
@@ -270,16 +254,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: exampleProfileTags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
 			require.Len(t, serverClient.appServers, 1)
 			appServer := serverClient.appServers[0]
@@ -324,16 +306,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: tags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
 
 			require.Len(t, serverClient.appServers, 1)
 			appServer := serverClient.appServers[0]
@@ -367,17 +347,14 @@ func TestRunAWSRolesAnywherProfileSyncer(t *testing.T) {
 			tags: tags,
 		}
 
-		synctest.Run(func() {
-			ctx, cancel := context.WithCancel(context.Background())
+		synctest.Test(t, func(t *testing.T) {
 			go func() {
-				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(ctx, params)
+				err := RunAWSRolesAnywhereProfileSyncerWhileLocked(t.Context(), params)
 				assert.NoError(t, err)
 			}()
 
 			// Wait for the 1st profile sync iteration.
 			synctest.Wait()
-			cancel()
-
 			status := serverClient.integrations[integrationWithProfileSync.GetName()].GetStatus()
 			require.NotNil(t, status)
 			lastSyncSummary := status.AWSRolesAnywhere.LastProfileSync
