+++ lib/srv/desktop/audit_compactor_test.go
@@ -1,5 +1,3 @@
-//go:build go1.24 && enablesynctest
-
 /*
  * Teleport
  * Copyright (C) 2025  Gravitational, Inc.
@@ -25,12 +23,13 @@ import (
 	"math"
 	"sync"
 	"testing"
-	"testing/synctest"
 	"time"
 
-	"github.com/gravitational/teleport/api/types/events"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/gravitational/teleport/api/types/events"
+	"github.com/gravitational/teleport/lib/utils/testutils/synctest"
 )
 
 func newReadEvent(path string, directory directoryID, offset uint64, length uint32) *events.DesktopSharedDirectoryRead {
@@ -69,8 +68,8 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("basic", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		ctx := t.Context()
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			// Read sequence A
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, 100))
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 100, 100))
@@ -96,7 +95,7 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("overflow", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
 			ctx := t.Context()
 			// Walk up to and beyond MaxUint32
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, math.MaxUint32-1))
@@ -113,7 +112,7 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("zero-length-event", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
 			ctx := t.Context()
 			// Create two read events with zero length, but with differing
 			// error codes. Neither should get compacted, as zero length
@@ -137,9 +136,8 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("complex", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		ctx := t.Context()
-		synctest.Run(func() {
-
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			// Three separate reads (with different lengths) of the same file
 			// Read sequence A
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, 100))
@@ -171,8 +169,8 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("expirations", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		ctx := t.Context()
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			// 2 sequential reads
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, 100))
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 100, 100))
@@ -220,8 +218,8 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("mix-reads-writes", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		ctx := t.Context()
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			// 3 sequential reads
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, 100))
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 100, 100))
@@ -240,8 +238,8 @@ func TestAuditCompactor(t *testing.T) {
 
 	t.Run("mix-files-and-directories", func(t *testing.T) {
 		auditEvents = auditEvents[:0]
-		ctx := t.Context()
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			// Identical offsets and lengths, but different path and/or directoryID
 			compactor.handleRead(ctx, newReadEvent("foo", 1, 0, 100))
 			compactor.handleRead(ctx, newReadEvent("foo", 2, 0, 100))
@@ -262,8 +260,8 @@ func TestAuditCompactor(t *testing.T) {
 	})
 
 	t.Run("racy-flush", func(t *testing.T) {
-		ctx := t.Context()
-		synctest.Run(func() {
+		synctest.Test(t, func(t *testing.T) {
+			ctx := t.Context()
 			auditEvents := make(chan events.AuditEvent)
 			compactor.emitFn = func(_ context.Context, ae events.AuditEvent) {
 				auditEvents <- ae
